% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manCTMed-boot-para-std-x-m-y.R
\name{BootParaStdXMY}
\alias{BootParaStdXMY}
\title{Parametric Bootstrap Confidence Intervals for X-M-Y
(Standardized)}
\usage{
BootParaStdXMY(boot, theta_hat, delta_t = 1:30, ncores = NULL)
}
\arguments{
\item{boot}{R object.
Output of the \code{\link[=BootPara]{BootPara()}} function.}

\item{theta_hat}{R object.
Output of the \code{\link[=ThetaHat]{ThetaHat()}} function.}

\item{delta_t}{Numeric vector.
Vector of time intervals.}

\item{ncores}{Positive integer.
Number of cores to use.}
}
\description{
The function generates parametric bootstrap method confidence intervals
for the mediation model \eqn{X \to M \to Y}
(Standardized).
}
\examples{
\dontrun{
set.seed(42)
library(dynr)
sim <- GenData(taskid = 1)
data <- RandomMeasurement(sim)
fit <- FitDynr(data, taskid = 1)
boot <- BootPara(
  fit = fit,
  path = getwd(),
  prefix = "pb",
  taskid = 1,
  B = 1000L
)
theta_hat <- ThetaHat(fit)
ci <- BootParaStdXMY(boot = boot, theta_hat = theta_hat)
plot(ci)
plot(ci, type = "bc")
}
}
\seealso{
Other Confidence Interval Functions:
\code{\link[=BootPara]{BootPara()}},
\code{\link[=BootParaStdXYM]{BootParaStdXYM()}},
\code{\link[=BootParaStdYMX]{BootParaStdYMX()}},
\code{\link[=BootParaXMY]{BootParaXMY()}},
\code{\link[=BootParaXYM]{BootParaXYM()}},
\code{\link[=BootParaYMX]{BootParaYMX()}},
\code{\link[=DeltaStdXMY]{DeltaStdXMY()}},
\code{\link[=DeltaStdXYM]{DeltaStdXYM()}},
\code{\link[=DeltaStdYMX]{DeltaStdYMX()}},
\code{\link[=DeltaXMY]{DeltaXMY()}},
\code{\link[=DeltaXYM]{DeltaXYM()}},
\code{\link[=DeltaYMX]{DeltaYMX()}},
\code{\link[=IllustrationBootPara]{IllustrationBootPara()}},
\code{\link[=MCStdXMY]{MCStdXMY()}},
\code{\link[=MCStdXYM]{MCStdXYM()}},
\code{\link[=MCStdYMX]{MCStdYMX()}},
\code{\link[=MCXMY]{MCXMY()}},
\code{\link[=MCXYM]{MCXYM()}},
\code{\link[=MCYMX]{MCYMX()}}
}
\concept{Confidence Interval Functions}
\keyword{ci}
\keyword{manCTMed}
