% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manCTMed-boot-para-x-y-m.R
\name{BootParaXYM}
\alias{BootParaXYM}
\title{Parametric Bootstrap Confidence Intervals for X-Y-M}
\usage{
BootParaXYM(boot, phi_hat, delta_t = 1:30, ncores = NULL)
}
\arguments{
\item{boot}{R object.
Output of the \code{\link[=BootPara]{BootPara()}} function.}

\item{phi_hat}{R object.
Output of the \code{\link[=PhiHat]{PhiHat()}} function.}

\item{delta_t}{Numeric vector.
Vector of time intervals.}

\item{ncores}{Positive integer.
Number of cores to use.}
}
\description{
The function generates parametric bootstrap method confidence intervals
for the mediation model \eqn{X \to Y \to M}.
}
\examples{
\dontrun{
set.seed(42)
library(dynr)
sim <- GenData(taskid = 1)
data <- RandomMeasurement(sim)
fit <- FitDynr(data, taskid = 1)
boot <- BootPara(
  fit = fit,
  path = getwd(),
  prefix = "pb",
  taskid = 1,
  B = 1000L
)
phi_hat <- PhiHat(fit)
ci <- BootParaXYM(boot = boot, phi_hat = phi_hat)
plot(ci)
plot(ci, type = "bc")
}
}
\seealso{
Other Confidence Interval Functions:
\code{\link[=BootPara]{BootPara()}},
\code{\link[=BootParaStdXMY]{BootParaStdXMY()}},
\code{\link[=BootParaStdXYM]{BootParaStdXYM()}},
\code{\link[=BootParaStdYMX]{BootParaStdYMX()}},
\code{\link[=BootParaXMY]{BootParaXMY()}},
\code{\link[=BootParaYMX]{BootParaYMX()}},
\code{\link[=DeltaStdXMY]{DeltaStdXMY()}},
\code{\link[=DeltaStdXYM]{DeltaStdXYM()}},
\code{\link[=DeltaStdYMX]{DeltaStdYMX()}},
\code{\link[=DeltaXMY]{DeltaXMY()}},
\code{\link[=DeltaXYM]{DeltaXYM()}},
\code{\link[=DeltaYMX]{DeltaYMX()}},
\code{\link[=IllustrationBootPara]{IllustrationBootPara()}},
\code{\link[=MCStdXMY]{MCStdXMY()}},
\code{\link[=MCStdXYM]{MCStdXYM()}},
\code{\link[=MCStdYMX]{MCStdYMX()}},
\code{\link[=MCXMY]{MCXMY()}},
\code{\link[=MCXYM]{MCXYM()}},
\code{\link[=MCYMX]{MCYMX()}}
}
\concept{Confidence Interval Functions}
\keyword{ci}
\keyword{manCTMed}
