% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manCTMed-ci.R
\name{CI}
\alias{CI}
\title{Generate Confidence Intervals
for the Total, Direct, and Indirect Effects}
\usage{
CI(x, delta_t = c(5, 10, 15, 20), R = 20000L)
}
\arguments{
\item{x}{R object.
Output of the \code{\link[=FitCtsem]{FitCtsem()}} or \code{\link[=FitDynr]{FitDynr()}} functions.}

\item{delta_t}{Numeric vector.
Vector of time-intervals to use.}

\item{R}{Positive integer.
Number of Monte Carlo replications.}
}
\description{
The function generates confidence intervals
for the total, direct, amd indirect effects
using the \link{cTMed} package.
}
\examples{
# dynr
library(dynr)
data <- Data(repid = 1, n = 50)
fit <- FitDynr(x = data)
CI(fit)

# ctsem
library(ctsem)
fit <- FitCtsem(x = data)
CI(fit)

}
\seealso{
Other Simulation Functions: 
\code{\link{Data}()},
\code{\link{FitCtsem}()},
\code{\link{FitDynr}()},
\code{\link{Replication}()},
\code{\link{Summarize}()}
}
\concept{Simulation Functions}
\keyword{manCTMed}
