% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manCTMed-collate.R
\name{Collate}
\alias{Collate}
\title{Collate Results}
\usage{
Collate(ns, reps, wd, ncores = NULL)
}
\arguments{
\item{ns}{A vector of sample sizes.}

\item{reps}{Replication numbers.}

\item{wd}{Working directory for output files.}

\item{ncores}{Positive integer.
Number of cores to use.}
}
\description{
Collate Results
}
\seealso{
Other Simulation Functions: 
\code{\link{CI}()},
\code{\link{Data}()},
\code{\link{FitDynr}()},
\code{\link{Plot}()},
\code{\link{Replication}()},
\code{\link{Summarize}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Simulation Functions}
\keyword{manCTMed}
