% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manCTMed-data.R
\name{Data}
\alias{Data}
\title{Simulate Data}
\usage{
Data(repid, n)
}
\arguments{
\item{repid}{Positive integer.
Replication ID.}

\item{n}{Positive integer.
Sample size.}
}
\description{
The function simulates data using
the \code{\link[simStateSpace:SimSSMOUFixed]{simStateSpace::SimSSMOUFixed()}} function.
}
\examples{
Data(repid = 1, n = 50)

}
\seealso{
Other Simulation Functions: 
\code{\link{CI}()},
\code{\link{Collate}()},
\code{\link{FitDynr}()},
\code{\link{Plot}()},
\code{\link{Replication}()},
\code{\link{Summarize}()}
}
\concept{Simulation Functions}
\keyword{manCTMed}
