% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manCTMed-fig-scatter-plot-coverage.R
\name{FigScatterPlotCoverage}
\alias{FigScatterPlotCoverage}
\title{Plot Coverage Probabilities}
\usage{
FigScatterPlotCoverage(results, delta_t = NULL, dynamics = 0, std = FALSE)
}
\arguments{
\item{results}{Summary results data frame.}

\item{delta_t}{Vector of time-interval value.
If \code{delta_t = NULL}, use all available time-intervals}

\item{dynamics}{Integer.
\code{dynamics = 0} for original drift matrix,
\code{dynamics = -1} for near-neutral dynamics, and
\code{dynamics = 1} for stronger damping.}

\item{std}{Logical.
If \code{std = TRUE}, standardized total, direct, and indirect effects.
If \code{std = FALSE}, unstandardized total, direct, and indirect effects.}
}
\description{
Coverage probabilities for the model \eqn{X \to M \to Y}.
}
\examples{
data(results, package = "manCTMed")
FigScatterPlotCoverage(results)
FigScatterPlotCoverage(results, delta_t = 1:14)
FigScatterPlotCoverage(results, delta_t = 15:30)

}
\seealso{
Other Figure Functions:
\code{\link[=FigPlotEffects]{FigPlotEffects()}},
\code{\link[=FigScatterPlotPower]{FigScatterPlotPower()}},
\code{\link[=FigScatterPlotSeBias]{FigScatterPlotSeBias()}},
\code{\link[=FigScatterPlotType1]{FigScatterPlotType1()}},
\code{\link[=IllustrationFigPlotEffects]{IllustrationFigPlotEffects()}},
\code{\link[=IllustrationFigScatterPlotCoverage]{IllustrationFigScatterPlotCoverage()}},
\code{\link[=IllustrationFigScatterPlotPower]{IllustrationFigScatterPlotPower()}},
\code{\link[=IllustrationFigScatterPlotSeBias]{IllustrationFigScatterPlotSeBias()}}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Figure Functions}
\keyword{figure}
\keyword{manCTMed}
