% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manCTMed-fit-ctsem.R
\name{FitCtsem}
\alias{FitCtsem}
\title{Fit the Model using the ctsem Package}
\usage{
FitCtsem(x)
}
\arguments{
\item{x}{R object.
Output of the \code{\link[=Data]{Data()}} function.}
}
\description{
The function fits the model using the \link{ctsem} package.
}
\examples{
library(ctsem)
data <- Data(repid = 1, n = 50)
fit <- FitCtsem(x = data)

}
\seealso{
Other Simulation Functions: 
\code{\link{CI}()},
\code{\link{Data}()},
\code{\link{FitDynr}()},
\code{\link{Replication}()},
\code{\link{Summarize}()}
}
\concept{Simulation Functions}
\keyword{manCTMed}
