% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manCTMed-fit-mx.R
\name{FitMx}
\alias{FitMx}
\title{Fit the Model using the OpenMx Package}
\usage{
FitMx(data, taskid)
}
\arguments{
\item{data}{R object.
Output of the \code{\link[=RandomMeasurement]{RandomMeasurement()}} function.}

\item{taskid}{Positive integer.
Task ID.}
}
\description{
The function fits the model using the \link[OpenMx:OpenMx]{OpenMx} package.
}
\examples{
\dontrun{
set.seed(42)
library(OpenMx)
sim <- GenData(taskid = 1)
data <- RandomMeasurement(sim)
fit <- FitMx(data, taskid = 1)
summary(fit)
}
}
\seealso{
Other Model Fitting Functions:
\code{\link[=FitDynr]{FitDynr()}},
\code{\link[=IllustrationFitDynr]{IllustrationFitDynr()}},
\code{\link[=IllustrationFitMx]{IllustrationFitMx()}},
\code{\link[=IllustrationMCPhiSigma]{IllustrationMCPhiSigma()}},
\code{\link[=IllustrationPrepData]{IllustrationPrepData()}},
\code{\link[=PhiHat]{PhiHat()}},
\code{\link[=ThetaHat]{ThetaHat()}}
}
\concept{Model Fitting Functions}
\keyword{fit}
\keyword{manCTMed}
