% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manCTMed-grundy2007.R
\docType{data}
\name{grundy2007}
\alias{grundy2007}
\title{grundy2007}
\format{
A dataframe with 4123 rows and 5 columns:

\describe{
\item{id}{
Participant ID variable.
}
\item{time}{
Time variable
}
\item{x}{
Mother-reported marital conflict.
}
\item{m}{
Maternal knowledge.
}
\item{y}{
Child-reported behavioral competence.
}
}
}
\usage{
data(grundy2007)
}
\description{
grundy2007
}
\references{
Grundy, A. M., Gondoli, D. M., & Blodgett Salafia, E. H. (2007).
Marital conflict and preadolescent behavioral competence:
Maternal knowledge as a longitudinal mediator.
Journal of Family Psychology, 21 (4), 675–682.
\doi{https://doi.org/10.1037/0893-3200.21.4.675}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{data}
\keyword{grundy}
