% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manCTMed-grundy2007-boot-para.R
\name{Grundy2007BootPara}
\alias{Grundy2007BootPara}
\title{Parametric Bootstrap (Grundy, et al. 2007)}
\usage{
Grundy2007BootPara(
  fit,
  data,
  path,
  prefix,
  B = 1000L,
  ncores = NULL,
  seed = NULL
)
}
\arguments{
\item{fit}{R object.
Output of the \code{\link[=FitDynr]{FitDynr()}} function.}

\item{data}{R object.
Output of the \code{\link[=Grundy2007GenData]{Grundy2007GenData()}} function.}

\item{path}{Path to a directory
to store bootstrap samples and estimates.}

\item{prefix}{Character string.
Prefix used for the file names
for the bootstrap samples and estimates.}

\item{B}{Positive integer.
Number of bootstrap samples.}

\item{ncores}{Positive integer.
Number of cores to use.}

\item{seed}{Integer.
Random seed.}
}
\description{
The function generates simulated datasets
based on a fitted model and refits the model
to each generated dataset using the \code{dynr} package.
}
\examples{
\dontrun{
library(dynr)
data <- Grundy2007GenData(taskid = 1)
fit <- Grundy2007FitDynr(data)
GrundyBootPara(
  fit = fit,
  data = data,
  path = getwd(),
  prefix = "pb",
  B = 1000L
)
}
}
\seealso{
Other Confidence Interval Functions: 
\code{\link{BootPara}()},
\code{\link{BootParaStdXMY}()},
\code{\link{BootParaStdXYM}()},
\code{\link{BootParaStdYMX}()},
\code{\link{BootParaXMY}()},
\code{\link{BootParaXYM}()},
\code{\link{BootParaYMX}()},
\code{\link{DeltaStdXMY}()},
\code{\link{DeltaStdXYM}()},
\code{\link{DeltaStdYMX}()},
\code{\link{DeltaXMY}()},
\code{\link{DeltaXYM}()},
\code{\link{DeltaYMX}()},
\code{\link{Grundy2007BootParaStdXMY}()},
\code{\link{Grundy2007BootParaStdXYM}()},
\code{\link{Grundy2007BootParaXMY}()},
\code{\link{Grundy2007BootParaXYM}()},
\code{\link{Grundy2007DeltaStdXMY}()},
\code{\link{Grundy2007DeltaStdXYM}()},
\code{\link{Grundy2007DeltaXMY}()},
\code{\link{Grundy2007DeltaXYM}()},
\code{\link{Grundy2007MCStdXMY}()},
\code{\link{Grundy2007MCStdXYM}()},
\code{\link{Grundy2007MCXMY}()},
\code{\link{Grundy2007MCXYM}()},
\code{\link{MCStdXMY}()},
\code{\link{MCStdXYM}()},
\code{\link{MCStdYMX}()},
\code{\link{MCXMY}()},
\code{\link{MCXYM}()},
\code{\link{MCYMX}()}
}
\concept{Confidence Interval Functions}
\keyword{ci}
\keyword{grundy}
\keyword{manCTMed}
