% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manCTMed-grundy2007-boot-para-x-y-m.R
\name{Grundy2007BootParaXYM}
\alias{Grundy2007BootParaXYM}
\title{Parametric Bootstrap Confidence Intervals for X-Y-M}
\usage{
Grundy2007BootParaXYM(boot, phi_hat, delta_t = 1:30, ncores = NULL)
}
\arguments{
\item{boot}{R object.
Output of the \code{\link[=BootPara]{BootPara()}} function.}

\item{phi_hat}{R object.
Output of the \code{\link[=PhiHat]{PhiHat()}} function.}

\item{delta_t}{Numeric vector.
Vector of time intervals.}

\item{ncores}{Positive integer.
Number of cores to use.}
}
\description{
The function generates parametric bootstrap method confidence intervals
for the mediation model \eqn{X \to Y \to M}.
}
\examples{
\dontrun{
library(dynr)
data <- Grundy2007GenData(taskid = 1)
fit <- Grundy2007FitDynr(data)
boot <- GrundyBootPara(
  fit = fit,
  data = data,
  path = getwd(),
  prefix = "pb",
  B = 1000L
)
phi_hat <- PhiHat(fit)
ci <- Grundy2007BootParaXYM(boot = boot, phi_hat = phi_hat)
plot(ci)
plot(ci, type = "bc")
}
}
\seealso{
Other Confidence Interval Functions: 
\code{\link{BootPara}()},
\code{\link{BootParaStdXMY}()},
\code{\link{BootParaStdXYM}()},
\code{\link{BootParaStdYMX}()},
\code{\link{BootParaXMY}()},
\code{\link{BootParaXYM}()},
\code{\link{BootParaYMX}()},
\code{\link{DeltaStdXMY}()},
\code{\link{DeltaStdXYM}()},
\code{\link{DeltaStdYMX}()},
\code{\link{DeltaXMY}()},
\code{\link{DeltaXYM}()},
\code{\link{DeltaYMX}()},
\code{\link{Grundy2007BootPara}()},
\code{\link{Grundy2007BootParaStdXMY}()},
\code{\link{Grundy2007BootParaStdXYM}()},
\code{\link{Grundy2007BootParaXMY}()},
\code{\link{Grundy2007DeltaStdXMY}()},
\code{\link{Grundy2007DeltaStdXYM}()},
\code{\link{Grundy2007DeltaXMY}()},
\code{\link{Grundy2007DeltaXYM}()},
\code{\link{Grundy2007MCStdXMY}()},
\code{\link{Grundy2007MCStdXYM}()},
\code{\link{Grundy2007MCXMY}()},
\code{\link{Grundy2007MCXYM}()},
\code{\link{MCStdXMY}()},
\code{\link{MCStdXYM}()},
\code{\link{MCStdYMX}()},
\code{\link{MCXMY}()},
\code{\link{MCXYM}()},
\code{\link{MCYMX}()}
}
\concept{Confidence Interval Functions}
\keyword{ci}
\keyword{grundy}
\keyword{manCTMed}
