% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manCTMed-grundy2007-delta-x-m-y.R
\name{Grundy2007DeltaXMY}
\alias{Grundy2007DeltaXMY}
\title{Delta Method Confidence Intervals for
Conflict-Knowledge-Competence (Grundy, et al. 2007)}
\usage{
Grundy2007DeltaXMY(phi_hat, delta_t = 1:30)
}
\arguments{
\item{phi_hat}{R object.
Output of the \code{\link[=PhiHat]{PhiHat()}} function.}

\item{delta_t}{Numeric vector.
Vector of time intervals.}
}
\description{
The function generates delta method confidence intervals
for the mediation model conflict-knowledge-competence.
}
\examples{
\dontrun{
set.seed(42)
library(dynr)
data <- Grundy2007GenData(taskid = 1)
fit <- Grundy2007FitDynr(data)
phi_hat <- PhiHat(fit)
ci <- Grundy2007DeltaXMY(phi_hat)
plot(ci)
}
}
\seealso{
Other Confidence Interval Functions: 
\code{\link{BootPara}()},
\code{\link{BootParaStdXMY}()},
\code{\link{BootParaStdXYM}()},
\code{\link{BootParaStdYMX}()},
\code{\link{BootParaXMY}()},
\code{\link{BootParaXYM}()},
\code{\link{BootParaYMX}()},
\code{\link{DeltaStdXMY}()},
\code{\link{DeltaStdXYM}()},
\code{\link{DeltaStdYMX}()},
\code{\link{DeltaXMY}()},
\code{\link{DeltaXYM}()},
\code{\link{DeltaYMX}()},
\code{\link{Grundy2007BootPara}()},
\code{\link{Grundy2007BootParaStdXMY}()},
\code{\link{Grundy2007BootParaStdXYM}()},
\code{\link{Grundy2007BootParaXMY}()},
\code{\link{Grundy2007BootParaXYM}()},
\code{\link{Grundy2007DeltaStdXMY}()},
\code{\link{Grundy2007DeltaStdXYM}()},
\code{\link{Grundy2007DeltaXYM}()},
\code{\link{Grundy2007MCStdXMY}()},
\code{\link{Grundy2007MCStdXYM}()},
\code{\link{Grundy2007MCXMY}()},
\code{\link{Grundy2007MCXYM}()},
\code{\link{MCStdXMY}()},
\code{\link{MCStdXYM}()},
\code{\link{MCStdYMX}()},
\code{\link{MCXMY}()},
\code{\link{MCXYM}()},
\code{\link{MCYMX}()}
}
\concept{Confidence Interval Functions}
\keyword{ci}
\keyword{grundy}
\keyword{manCTMed}
