% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manCTMed-grundy2007-fit-dynr.R
\name{Grundy2007FitDynr}
\alias{Grundy2007FitDynr}
\title{Fit the CT-VAR Model (Grundy, et al. 2007)}
\usage{
Grundy2007FitDynr(data)
}
\arguments{
\item{data}{R object.
Output of the \code{\link[=Grundy2007GenData]{Grundy2007GenData()}} function.}
}
\description{
Fit the CT-VAR Model (Grundy, et al. 2007)
}
\examples{
\dontrun{
library(dynr)
data <- Grundy2007GenData(taskid = 1)
fit <- Grundy2007FitDynr(data)
summary(fit)
}
}
\seealso{
Other Model Fitting Functions: 
\code{\link{FitDynr}()},
\code{\link{FitMx}()},
\code{\link{PhiHat}()},
\code{\link{ThetaHat}()}
}
\concept{Model Fitting Functions}
\keyword{fit}
\keyword{grundy}
\keyword{manCTMed}
