% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manCTMed-grundy2007-gen-data.R
\name{Grundy2007GenData}
\alias{Grundy2007GenData}
\title{Generate Data (Grundy, et al. 2007)}
\usage{
Grundy2007GenData(taskid, seed = NULL, empirical = FALSE)
}
\arguments{
\item{taskid}{Positive integer.
Task ID.}

\item{seed}{Integer.
Random seed.}

\item{empirical}{logical. If true, mu and Sigma specify the empirical
    not population mean and covariance matrix.}
}
\description{
Generate Data (Grundy, et al. 2007)
}
\examples{
\dontrun{
Grundy2007GenData(taskid = 1)
}
}
\seealso{
Other Data Generation Functions: 
\code{\link{GenData}()},
\code{\link{RandomMeasurement}()}
}
\concept{Data Generation Functions}
\keyword{data}
\keyword{grundy}
\keyword{manCTMed}
