% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manCTMed-illustration-data-dist.R
\docType{data}
\name{illustration_dist}
\alias{illustration_dist}
\title{Illustration Sampling Distribution}
\format{
A matrix with 1000 rows and 27 columns:

\describe{
\item{phi_xx}{
Elements of the drift matrix.
}
\item{sigma_xx}{
Elements of the process noise covariance matrix.
}
\item{theta_xx}{
Elements of the measurement error covariance matrix.
}
\item{mu0_x}{
Elements of the initial condition mean vector.
}
\item{sigma0_xx}{
Elements of the initial condition covariance matrix.
}
}
}
\usage{
data(illustration_dist)
}
\description{
Illustration Sampling Distribution
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{data}
\keyword{illustration}
