% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manCTMed-illustration-data-dist-dt-mc.R
\docType{data}
\name{illustration_dist_dt_mc}
\alias{illustration_dist_dt_mc}
\title{Illustration Sampling Distribution
Discrete Time - Time Interval of 1
(Monte Carlo Method)}
\format{
A matrix with 20000 rows and 15 columns:

\describe{
\item{beta_xx}{
Elements of the matrix of lagged coefficients.
}
\item{psi_xx}{
Elements of the process noise covariance matrix.
}
}
}
\usage{
data(illustration_dist_dt_mc)
}
\description{
Illustration Sampling Distribution
Discrete Time - Time Interval of 1
(Monte Carlo Method)
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{data}
\keyword{illustration}
