% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manCTMed-illustration-data-dist-mc.R
\docType{data}
\name{illustration_dist_mc}
\alias{illustration_dist_mc}
\title{Illustration Sampling Distribution (Monte Carlo Method)}
\format{
A matrix with 20000 rows and 15 columns:

\describe{
\item{phi_xx}{
Elements of the drift matrix.
}
\item{sigma_xx}{
Elements of the process noise covariance matrix.
}
}
}
\usage{
data(illustration_dist_mc)
}
\description{
Illustration Sampling Distribution (Monte Carlo Method)
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{data}
\keyword{illustration}
