% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manCTMed-illustration-data-dist-med.R
\docType{data}
\name{illustration_dist_med}
\alias{illustration_dist_med}
\title{Illustration Sampling Distribution
Total, Direct, and Indirect Effects - Time Interval of 1}
\format{
A matrix with 1000 rows and 27 columns:

\describe{
\item{total}{
Total effect.
}
\item{direct}{
Direct effect.
}
\item{indirect}{
Indirect effect.
}
}
}
\usage{
data(illustration_dist_med)
}
\description{
Illustration Sampling Distribution
Total, Direct, and Indirect Effects - Time Interval of 1
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{data}
\keyword{illustration}
