% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manCTMed-illustration-boot-para.R
\name{IllustrationBootPara}
\alias{IllustrationBootPara}
\title{Parametric Bootstrap (Illustration)}
\usage{
IllustrationBootPara(
  fit,
  path,
  prefix,
  taskid,
  B = 1000L,
  ncores = NULL,
  seed = NULL
)
}
\arguments{
\item{fit}{R object.
Fitted CT-VAR model.}

\item{path}{Path to a directory
to store bootstrap samples and estimates.}

\item{prefix}{Character string.
Prefix used for the file names
for the bootstrap samples and estimates.}

\item{taskid}{Positive integer.
Task ID.}

\item{B}{Positive integer.
Number of bootstrap samples.}

\item{ncores}{Positive integer.
Number of cores to use.}

\item{seed}{Integer.
Random seed.}
}
\description{
The function generates simulated datasets
based on a fitted model and refits the model
to each generated dataset using the \code{dynr} package.
}
\examples{
\dontrun{
library(dynr)
sim <- IllustrationGenData(seed = 42)
data <- IllustrationPrepData(sim)
fit <- IllustrationFitDynr(data)
summary(fit)
IllustrationBootPara(
  fit = fit,
  path = getwd(),
  prefix = "pb",
  taskid = 1,
  B = 1000L,
  seed = 42
)
}
}
\seealso{
Other Confidence Interval Functions:
\code{\link[=BootPara]{BootPara()}},
\code{\link[=BootParaStdXMY]{BootParaStdXMY()}},
\code{\link[=BootParaStdXYM]{BootParaStdXYM()}},
\code{\link[=BootParaStdYMX]{BootParaStdYMX()}},
\code{\link[=BootParaXMY]{BootParaXMY()}},
\code{\link[=BootParaXYM]{BootParaXYM()}},
\code{\link[=BootParaYMX]{BootParaYMX()}},
\code{\link[=DeltaStdXMY]{DeltaStdXMY()}},
\code{\link[=DeltaStdXYM]{DeltaStdXYM()}},
\code{\link[=DeltaStdYMX]{DeltaStdYMX()}},
\code{\link[=DeltaXMY]{DeltaXMY()}},
\code{\link[=DeltaXYM]{DeltaXYM()}},
\code{\link[=DeltaYMX]{DeltaYMX()}},
\code{\link[=MCStdXMY]{MCStdXMY()}},
\code{\link[=MCStdXYM]{MCStdXYM()}},
\code{\link[=MCStdYMX]{MCStdYMX()}},
\code{\link[=MCXMY]{MCXMY()}},
\code{\link[=MCXYM]{MCXYM()}},
\code{\link[=MCYMX]{MCYMX()}}
}
\concept{Confidence Interval Functions}
\keyword{ci}
\keyword{illustration}
\keyword{manCTMed}
