% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manCTMed-illustration-compress.R
\name{IllustrationCompress}
\alias{IllustrationCompress}
\title{Compress Replication (Illustration)}
\usage{
IllustrationCompress(taskid, repid, output_folder)
}
\arguments{
\item{taskid}{Positive integer.
Task ID.}

\item{repid}{Positive integer.
Replication ID.}

\item{output_folder}{Character string.
Output folder.}
}
\value{
The output is saved as an external file in \code{output_folder}.
}
\description{
Compress Replication (Illustration)
}
\seealso{
Other Compression Functions:
\code{\link[=Compress]{Compress()}}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Compression Functions}
\keyword{compress}
\keyword{manCTMed}
