% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manCTMed-illustration-fig-scatter-plot-power.R
\name{IllustrationFigScatterPlotPower}
\alias{IllustrationFigScatterPlotPower}
\title{Illustration Plot Statistical Power}
\usage{
IllustrationFigScatterPlotPower(illustration_results)
}
\arguments{
\item{illustration_results}{Summary results data frame.}
}
\description{
Statistical Power for the model \eqn{X \to M \to Y}.
}
\examples{
data(illustration_results, package = "manCTMed")
IllustrationFigScatterPlotPower(illustration_results)

}
\seealso{
Other Figure Functions:
\code{\link[=FigPlotEffects]{FigPlotEffects()}},
\code{\link[=FigScatterPlotCoverage]{FigScatterPlotCoverage()}},
\code{\link[=FigScatterPlotPower]{FigScatterPlotPower()}},
\code{\link[=FigScatterPlotSeBias]{FigScatterPlotSeBias()}},
\code{\link[=FigScatterPlotType1]{FigScatterPlotType1()}},
\code{\link[=IllustrationFigPlotEffects]{IllustrationFigPlotEffects()}},
\code{\link[=IllustrationFigScatterPlotCoverage]{IllustrationFigScatterPlotCoverage()}},
\code{\link[=IllustrationFigScatterPlotSeBias]{IllustrationFigScatterPlotSeBias()}}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Figure Functions}
\keyword{figure}
\keyword{illustration}
\keyword{manCTMed}
