% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manCTMed-illustration-fit-dynr.R
\name{IllustrationFitDynr}
\alias{IllustrationFitDynr}
\title{Fit the Model using the dynr Package (Illustration)}
\usage{
IllustrationFitDynr(data)
}
\arguments{
\item{data}{R object.
Output of the \code{\link[=IllustrationPrepData]{IllustrationPrepData()}} function.}
}
\description{
The function fits the model using the \link[dynr:dynr]{dynr} package.
}
\examples{
\dontrun{
library(dynr)
sim <- IllustrationGenData(seed = 42)
data <- IllustrationPrepData(sim)
fit <- IllustrationFitDynr(data)
summary(fit)
}
}
\seealso{
Other Model Fitting Functions:
\code{\link[=FitDynr]{FitDynr()}},
\code{\link[=FitMx]{FitMx()}},
\code{\link[=IllustrationFitMx]{IllustrationFitMx()}},
\code{\link[=IllustrationMCPhiSigma]{IllustrationMCPhiSigma()}},
\code{\link[=IllustrationPrepData]{IllustrationPrepData()}},
\code{\link[=PhiHat]{PhiHat()}},
\code{\link[=ThetaHat]{ThetaHat()}}
}
\concept{Model Fitting Functions}
\keyword{fit}
\keyword{illustration}
\keyword{manCTMed}
