% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manCTMed-illustration-fit-mx.R
\name{IllustrationFitMx}
\alias{IllustrationFitMx}
\title{Fit the Model using the OpenMx Package  (Illustration)}
\usage{
IllustrationFitMx(data)
}
\arguments{
\item{data}{R object.
Output of the \code{\link[=IllustrationPrepData]{IllustrationPrepData()}} function.}
}
\description{
The function fits the model using the \link[OpenMx:OpenMx]{OpenMx} package.
}
\examples{
\dontrun{
library(OpenMx)
sim <- IllustrationGenData(seed = 42)
data <- IllustrationPrepData(sim)
fit <- IllustrationFitMx(data)
summary(fit)
}
}
\seealso{
Other Model Fitting Functions:
\code{\link[=FitDynr]{FitDynr()}},
\code{\link[=FitMx]{FitMx()}},
\code{\link[=IllustrationFitDynr]{IllustrationFitDynr()}},
\code{\link[=IllustrationMCPhiSigma]{IllustrationMCPhiSigma()}},
\code{\link[=IllustrationPrepData]{IllustrationPrepData()}},
\code{\link[=PhiHat]{PhiHat()}},
\code{\link[=ThetaHat]{ThetaHat()}}
}
\concept{Model Fitting Functions}
\keyword{fit}
\keyword{illustration}
\keyword{manCTMed}
