% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manCTMed-illustration-prep-data.R
\name{IllustrationPrepData}
\alias{IllustrationPrepData}
\title{Prepare Data Before Model Fitting (Illustration)}
\usage{
IllustrationPrepData(sim)
}
\arguments{
\item{sim}{R object.
Output of the \code{\link[=IllustrationGenData]{IllustrationGenData()}} function.}
}
\description{
The function converts the output of \code{\link[=IllustrationGenData]{IllustrationGenData()}}
into a data frame.
}
\examples{
\dontrun{
sim <- IllustrationGenData(seed = 42)
data <- IllustrationPrepData(sim)
head(data)
dim(data)
}
}
\seealso{
Other Model Fitting Functions:
\code{\link[=FitDynr]{FitDynr()}},
\code{\link[=FitMx]{FitMx()}},
\code{\link[=IllustrationFitDynr]{IllustrationFitDynr()}},
\code{\link[=IllustrationFitMx]{IllustrationFitMx()}},
\code{\link[=IllustrationMCPhiSigma]{IllustrationMCPhiSigma()}},
\code{\link[=PhiHat]{PhiHat()}},
\code{\link[=ThetaHat]{ThetaHat()}}
}
\concept{Model Fitting Functions}
\keyword{gendata}
\keyword{illustration}
\keyword{manCTMed}
