% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manCTMed-illustration-sum-hit.R
\name{IllustrationSumHit}
\alias{IllustrationSumHit}
\title{Summarize Simulations - CI Hit}
\usage{
IllustrationSumHit(taskid, reps, output_folder, output_type, ncores)
}
\arguments{
\item{taskid}{Positive integer.
Task ID.}

\item{reps}{Positive integer.
Number of replications.}

\item{output_folder}{Character string.
Output folder.}

\item{output_type}{Character string.
Output type.}

\item{ncores}{Positive integer.
Number of cores to use.}
}
\value{
Returns a numeric matrix.
}
\description{
Summarize Simulations - CI Hit
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{ci}
\keyword{manCTMed}
\keyword{summary}
