% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manCTMed-package.R
\docType{package}
\name{manCTMed-package}
\alias{manCTMed}
\alias{manCTMed-package}
\title{manCTMed: Continuous Time Mediation}
\description{
Research compendium for the manuscript Pesigan, I. J. A., Russell, M. A., & Chow, S.-M. (2025). Inferences and Effect Sizes for Direct, Indirect, and Total Effects in Continuous-Time Mediation Models. Psychological Methods. \doi{10.1037/met0000779}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jeksterslab/manCTMed}
  \item \url{https://jeksterslab.github.io/manCTMed/}
  \item \url{https://osf.io/qwnmf/}
  \item \doi{10.1037/met0000779}
  \item Report bugs at \url{https://github.com/jeksterslab/manCTMed/issues}
}

}
\author{
\strong{Maintainer}: Ivan Jacob Agaloos Pesigan \email{r.jeksterslab@gmail.com} (\href{https://orcid.org/0000-0003-4818-8420}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Ivan Jacob Agaloos Pesigan \email{r.jeksterslab@gmail.com} (\href{https://orcid.org/0000-0003-4818-8420}{ORCID}) [copyright holder]
}

Other contributors:
\itemize{
  \item Michael A. Russell \email{mar60@psu.edu} (\href{https://orcid.org/0000-0002-3956-604X}{ORCID}) [contributor]
  \item Sy-Miin Chow \email{quc16@psu.edu} (\href{https://orcid.org/0000-0003-1938-027X}{ORCID}) [contributor]
}

}
\keyword{internal}
