% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manCTMed-mc-x-m-y.R
\name{MCXMY}
\alias{MCXMY}
\title{Monte Carlo Method Confidence Intervals for X-M-Y}
\usage{
MCXMY(phi_hat, delta_t = 1:30, R = 20000L, seed = NULL)
}
\arguments{
\item{phi_hat}{R object.
Output of the \code{\link[=PhiHat]{PhiHat()}} function.}

\item{delta_t}{Numeric vector.
Vector of time intervals.}

\item{R}{Positive integer.
Number of Monte Carlo replications.}

\item{seed}{Integer.
Random seed.}
}
\description{
The function generates Monte Carlo method confidence intervals
for the mediation model \eqn{X \to M \to Y}.
}
\examples{
\dontrun{
set.seed(42)
library(dynr)
sim <- GenData(taskid = 1)
data <- RandomMeasurement(sim)
fit <- FitDynr(data, taskid = 1)
phi_hat <- PhiHat(fit)
ci <- MCXMY(phi_hat, seed = 42)
plot(ci)
}
}
\seealso{
Other Confidence Interval Functions:
\code{\link[=BootPara]{BootPara()}},
\code{\link[=BootParaStdXMY]{BootParaStdXMY()}},
\code{\link[=BootParaStdXYM]{BootParaStdXYM()}},
\code{\link[=BootParaStdYMX]{BootParaStdYMX()}},
\code{\link[=BootParaXMY]{BootParaXMY()}},
\code{\link[=BootParaXYM]{BootParaXYM()}},
\code{\link[=BootParaYMX]{BootParaYMX()}},
\code{\link[=DeltaStdXMY]{DeltaStdXMY()}},
\code{\link[=DeltaStdXYM]{DeltaStdXYM()}},
\code{\link[=DeltaStdYMX]{DeltaStdYMX()}},
\code{\link[=DeltaXMY]{DeltaXMY()}},
\code{\link[=DeltaXYM]{DeltaXYM()}},
\code{\link[=DeltaYMX]{DeltaYMX()}},
\code{\link[=IllustrationBootPara]{IllustrationBootPara()}},
\code{\link[=MCStdXMY]{MCStdXMY()}},
\code{\link[=MCStdXYM]{MCStdXYM()}},
\code{\link[=MCStdYMX]{MCStdYMX()}},
\code{\link[=MCXYM]{MCXYM()}},
\code{\link[=MCYMX]{MCYMX()}}
}
\concept{Confidence Interval Functions}
\keyword{ci}
\keyword{manCTMed}
