% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manCTMed-plot.R
\name{Plot}
\alias{Plot}
\title{Plot Results}
\usage{
Plot(x)
}
\arguments{
\item{x}{R object.
Output of the \code{\link[=Summarize]{Summarize()}} function.}
}
\description{
Plot Results
}
\seealso{
Other Simulation Functions: 
\code{\link{CI}()},
\code{\link{Collate}()},
\code{\link{Data}()},
\code{\link{FitDynr}()},
\code{\link{Replication}()},
\code{\link{Summarize}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Simulation Functions}
\keyword{manCTMed}
