% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manCTMed-plot-coverage.R
\name{PlotCoverage}
\alias{PlotCoverage}
\title{Plot Coverage Probabilities}
\usage{
PlotCoverage(x)
}
\arguments{
\item{x}{Summary results data frame.}
}
\description{
Plot Coverage Probabilities
}
\seealso{
Other Simulation Functions: 
\code{\link{PlotPower}()},
\code{\link{PlotType1}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Simulation Functions}
\keyword{manCTMed}
