% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manCTMed-plot-power.R
\name{PlotPower}
\alias{PlotPower}
\title{Plot Statistical Power}
\usage{
PlotPower(x)
}
\arguments{
\item{x}{Summary results data frame.}
}
\description{
Plot Statistical Power
}
\seealso{
Other Figure Functions: 
\code{\link{PlotCoverage}()},
\code{\link{PlotType1}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Figure Functions}
\keyword{manCTMed}
\keyword{plot}
