% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manCTMed-plot-type-1.R,
%   R/manCTMed-plot-type-2.R
\name{PlotType1}
\alias{PlotType1}
\title{Plot Type I Error}
\usage{
PlotType1(x)

PlotType1(x)
}
\arguments{
\item{x}{R object.
Output of the \code{\link[=Summarize]{Summarize()}} function.}
}
\description{
Plot Type I Error

Plot Type II Error
}
\seealso{
Other Simulation Functions: 
\code{\link{CI}()},
\code{\link{Collate}()},
\code{\link{Data}()},
\code{\link{FitDynr}()},
\code{\link{PlotCoverage}()},
\code{\link{Replication}()},
\code{\link{Summarize}()}

Other Simulation Functions: 
\code{\link{CI}()},
\code{\link{Collate}()},
\code{\link{Data}()},
\code{\link{FitDynr}()},
\code{\link{PlotCoverage}()},
\code{\link{Replication}()},
\code{\link{Summarize}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Simulation Functions}
\keyword{manCTMed}
