% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manCTMed-random-measurement.R
\name{RandomMeasurement}
\alias{RandomMeasurement}
\title{Simulate Random Measurement}
\usage{
RandomMeasurement(sim)
}
\arguments{
\item{sim}{R object.
Output of the \code{\link[=GenData]{GenData()}} function.}
}
\description{
The function randomly selects 100 observations from the generated data
and replaces the unselected observations with \code{NA}.
}
\examples{
\dontrun{
set.seed(42)
sim <- GenData(taskid = 1)
RandomMeasurement(sim)
}
}
\seealso{
Other Data Generation Functions:
\code{\link[=GenData]{GenData()}},
\code{\link[=IllustrationGenData]{IllustrationGenData()}}
}
\concept{Data Generation Functions}
\keyword{gendata}
\keyword{manCTMed}
