% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manCTMed-replication.R
\name{Replication}
\alias{Replication}
\title{Simulation Replication}
\usage{
Replication(i, n)
}
\arguments{
\item{i}{Positive integer.
Replication ID.}

\item{n}{Positive integer.
Sample size.}
}
\description{
Performs a single simulation replication involving the following:
\itemize{
\item Data generation
\item Model fitting using the dynr package
\item Model fitting using the ctsem package
\item Delta method using the \code{\link[cTMed:DeltaMed]{cTMed::DeltaMed()}} function
\item Monte Carlo method using the \code{\link[cTMed:MCMed]{cTMed::MCMed()}} function
\item Posterior distribution using the \code{\link[cTMed:PosteriorMed]{cTMed::PosteriorMed()}} function
}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Simulation Functions}
\keyword{manCTMed}
