% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSets-ryan2021.R
\docType{data}
\name{ryan2021}
\alias{ryan2021}
\title{Experience Sampling Data (n = 1) used as Empirical Example in
Ryan and Hamaker (2021)}
\format{
The data set has the following variables:
\describe{
\item{id}{ID}
\item{time}{Time variable}
\item{s}{Self-doubt}
\item{f}{Fatigue}
\item{i}{Irritable}
\item{r}{Restless}
}
}
\usage{
ryan2021
}
\description{
Experience Sampling Data (n = 1) used as Empirical Example in
Ryan and Hamaker (2021)
}
\references{
Kossakowski, J., Groot, P., Haslbeck, J. M., Borsboom, D., & Wichers, M.
(2017).
Data from
"Critical slowing down
as a personalized early warning signal for depression".
\emph{Journal of Open Psychology Data},
\emph{5}(1).
\doi{10.5334/jopd.29}

Ryan, O., & Hamaker, E. L. (2021).
Time to intervene:
A continuous-time approach to network analysis and centrality.
\emph{Psychometrika},
\emph{87} (1), 214–252.
\doi{10.1007/s11336-021-09767-0}
}
\keyword{data}
