% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manCTMed-ryan2021phi.R
\docType{data}
\name{ryan2021phi}
\alias{ryan2021phi}
\title{Simulation Parameters}
\format{
A list with the following elements:

\describe{
\item{phi}{
The estimated drift matrix.
}
\item{vcov}{
The estimated sampling variance-covariance matrix
of the estimated drift matrix.
}
}
}
\usage{
ryan2021phi
}
\description{
Simulation Parameters
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{data}
\keyword{empirical}
