% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manCTMed-sim-grundy2007-gen-data.R
\name{SimGrundy2007GenData}
\alias{SimGrundy2007GenData}
\title{Simulation Replication - Grundy2007GenData}
\usage{
SimGrundy2007GenData(
  taskid,
  repid,
  output_folder,
  seed,
  suffix,
  overwrite,
  integrity
)
}
\arguments{
\item{taskid}{Positive integer.
Task ID.}

\item{repid}{Positive integer.
Replication ID.}

\item{output_folder}{Character string.
Output folder.}

\item{seed}{Integer.
Random seed.}

\item{suffix}{Character string.
Output of \code{manCTMed:::.SimSuffix()}.}

\item{overwrite}{Logical.
Overwrite existing output in \code{output_folder}.}

\item{integrity}{Logical.
If \code{integrity = TRUE},
check for the output file integrity when \code{overwrite = FALSE}.}
}
\value{
The output is saved as an external file in \code{output_folder}.
}
\description{
Simulation Replication - Grundy2007GenData
}
\details{
This function is executed via the \code{SimGrundy2007} function.
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{Grundy2007GenData}
\keyword{manCTMed}
\keyword{simulation}
