% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manCTMed-illustration-sim.R
\name{SimIllustration}
\alias{SimIllustration}
\title{Simulation Replication (Illustration)}
\usage{
SimIllustration(
  taskid,
  repid,
  output_folder,
  overwrite,
  integrity,
  seed,
  ci,
  pb,
  delta_t,
  R,
  B
)
}
\arguments{
\item{taskid}{Positive integer.
Task ID.}

\item{repid}{Positive integer.
Replication ID.}

\item{output_folder}{Character string.
Output folder.}

\item{overwrite}{Logical.
Overwrite existing output in \code{output_folder}.}

\item{integrity}{Logical.
If \code{integrity = TRUE},
check for the output file integrity when \code{overwrite = FALSE}.}

\item{seed}{Integer.
Random seed.}

\item{ci}{Logical.
Run simulations for confidence intervals.}

\item{pb}{Logical.
Run simulations for parametric bootstrap confidence intervals.}

\item{delta_t}{Numeric vector.
Vector of time intervals.}

\item{R}{Positive integer.
Number of Monte Carlo replications.}

\item{B}{Positive integer.
Number of bootstrap samples.}
}
\value{
The output is saved as an external file in \code{output_folder}.
}
\description{
Simulation Replication (Illustration)
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{illustration}
\keyword{manCTMed}
