% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manCTMed-sim-para.R
\name{SimPara}
\alias{SimPara}
\title{Simulation Replication Parametric Bootstrap (Parallel)}
\usage{
SimPara(taskid, repid, output_folder, overwrite, integrity, seed, B)
}
\arguments{
\item{taskid}{Positive integer.
Task ID.}

\item{repid}{Positive integer.
Replication ID.}

\item{output_folder}{Character string.
Output folder.}

\item{overwrite}{Logical.
Overwrite existing output in \code{output_folder}.}

\item{integrity}{Logical.
If \code{integrity = TRUE},
check for the output file integrity when \code{overwrite = FALSE}.}

\item{seed}{Integer.
Random seed.}

\item{B}{Positive integer.
Number of bootstrap samples.}
}
\value{
The output is saved as an external file in \code{output_folder}.
}
\description{
Simulation Replication Parametric Bootstrap (Parallel)
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{manCTMed}
\keyword{simulation}
