% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manCTMed-sum-dynr-delta-std-x-m-y.R
\name{SumDynrDeltaStdXMY}
\alias{SumDynrDeltaStdXMY}
\title{Summary (DynrDeltaStdXMY)}
\usage{
SumDynrDeltaStdXMY(taskid, reps, output_folder, overwrite, integrity)
}
\arguments{
\item{taskid}{Positive integer.
Task ID.}

\item{reps}{Positive integer.
Number of replications.}

\item{output_folder}{Character string.
Output folder.}

\item{overwrite}{Logical.
Overwrite existing output in \code{output_folder}.}

\item{integrity}{Logical.
If \code{integrity = TRUE},
check for the output file integrity when \code{overwrite = FALSE}.}
}
\value{
The output is saved as an external file in \code{output_folder}.
}
\description{
Summary (DynrDeltaStdXMY)
}
\details{
This function is executed via the \code{Sum} function.
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{manCTMed}
\keyword{simulation}
\keyword{summary}
