% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manCTMed-sum-grundy2007-hit.R
\name{SumGrundy2007Hit}
\alias{SumGrundy2007Hit}
\title{Summarize Simulations - CI Hit
(Grundy, et al. 2007)}
\usage{
SumGrundy2007Hit(taskid, reps, output_folder, output_type, ncores)
}
\arguments{
\item{taskid}{Positive integer.
Task ID.}

\item{reps}{Positive integer.
Number of replications.}

\item{output_folder}{Character string.
Output folder.}

\item{output_type}{Character string.
Output type.
Valid values include
\code{"data"},
\code{"fit-dynr"},
\code{"dynr-delta-xmy"},
\code{"dynr-delta-ymx"},
\code{"dynr-mc-xmy"}, and
\code{"dynr-mc-ymx"}}

\item{ncores}{Positive integer.
Number of cores to use.}
}
\value{
Returns a numeric matrix.
}
\description{
Summarize Simulations - CI Hit
(Grundy, et al. 2007)
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{ci}
\keyword{manCTMed}
\keyword{summary}
