% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manCTMed-boot-para.R
\name{BootPara}
\alias{BootPara}
\title{Parametric Bootstrap}
\usage{
BootPara(
  fit,
  path,
  prefix,
  taskid,
  B = 1000L,
  ncores = NULL,
  seed = NULL,
  clean = TRUE
)
}
\arguments{
\item{fit}{R object.
Output of the \code{\link[=FitDynr]{FitDynr()}}, \code{\link[=FitMx]{FitMx()}},
\code{\link[=IllustrationFitDynr]{IllustrationFitDynr()}}, or \code{\link[=IllustrationFitMx]{IllustrationFitMx()}},
functions.}

\item{path}{Path to a directory
to store bootstrap samples and estimates.}

\item{prefix}{Character string.
Prefix used for the file names
for the bootstrap samples and estimates.}

\item{taskid}{Positive integer.
Task ID.}

\item{B}{Positive integer.
Number of bootstrap samples.}

\item{ncores}{Positive integer.
Number of cores to use.}

\item{seed}{Integer.
Random seed.}

\item{clean}{Logical.
If \code{clean = TRUE},
delete intermediate files generated by the function.}
}
\description{
The function generates simulated datasets
based on a fitted model and refits the model
to each generated dataset using the \code{dynr} package.
}
\examples{
\dontrun{
set.seed(42)
library(dynr)
sim <- GenData(taskid = 1)
data <- RandomMeasurement(sim)
fit <- FitDynr(data, taskid = 1)
BootPara(
  fit = fit,
  path = getwd(),
  prefix = "pb",
  taskid = 1,
  B = 1000L
)
}
}
\seealso{
Other Confidence Interval Functions: 
\code{\link{BootParaStdXMY}()},
\code{\link{BootParaStdXYM}()},
\code{\link{BootParaStdYMX}()},
\code{\link{BootParaXMY}()},
\code{\link{BootParaXYM}()},
\code{\link{BootParaYMX}()},
\code{\link{DeltaStdXMY}()},
\code{\link{DeltaStdXYM}()},
\code{\link{DeltaStdYMX}()},
\code{\link{DeltaXMY}()},
\code{\link{DeltaXYM}()},
\code{\link{DeltaYMX}()},
\code{\link{IllustrationBootPara}()},
\code{\link{MCStdXMY}()},
\code{\link{MCStdXYM}()},
\code{\link{MCStdYMX}()},
\code{\link{MCXMY}()},
\code{\link{MCXYM}()},
\code{\link{MCYMX}()}
}
\concept{Confidence Interval Functions}
\keyword{ci}
\keyword{manCTMed}
