% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manCTMed-boot-para-y-m-x.R
\name{BootParaYMX}
\alias{BootParaYMX}
\title{Parametric Bootstrap Confidence Intervals for Y-M-X}
\usage{
BootParaYMX(boot, phi_hat, delta_t = 1:30, ncores = NULL)
}
\arguments{
\item{boot}{R object.
Output of the \code{\link[=BootPara]{BootPara()}} function.}

\item{phi_hat}{R object.
Output of the \code{\link[=PhiHat]{PhiHat()}} function.}

\item{delta_t}{Numeric vector.
Vector of time intervals.}

\item{ncores}{Positive integer.
Number of cores to use.}
}
\description{
The function generates parametric bootstrap method confidence intervals
for the mediation model \eqn{Y \to M \to X}.
}
\examples{
\dontrun{
set.seed(42)
library(dynr)
sim <- GenData(taskid = 1)
data <- RandomMeasurement(sim)
fit <- FitDynr(data, taskid = 1)
boot <- BootPara(
  fit = fit,
  path = getwd(),
  prefix = "pb",
  taskid = 1,
  B = 1000L
)
phi_hat <- PhiHat(fit)
ci <- BootParaYMX(boot = boot, phi_hat = phi_hat)
summary(ci)
summary(ci, type = "bc")
}
}
\seealso{
Other Confidence Interval Functions: 
\code{\link{BootPara}()},
\code{\link{BootParaStdXMY}()},
\code{\link{BootParaStdYMX}()},
\code{\link{BootParaXMY}()},
\code{\link{DeltaStdXMY}()},
\code{\link{DeltaStdYMX}()},
\code{\link{DeltaXMY}()},
\code{\link{DeltaYMX}()},
\code{\link{MCStdXMY}()},
\code{\link{MCStdYMX}()},
\code{\link{MCXMY}()},
\code{\link{MCYMX}()}
}
\concept{Confidence Interval Functions}
\keyword{ci}
\keyword{manCTMed}
