% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manCTMed-delta-std-y-m-x.R
\name{DeltaStdYMX}
\alias{DeltaStdYMX}
\title{Delta Method Confidence Intervals for Y-M-X (Standardized)}
\usage{
DeltaStdYMX(theta_hat, delta_t = 1:30)
}
\arguments{
\item{theta_hat}{R object.
Output of the \code{\link[=ThetaHat]{ThetaHat()}} function.}

\item{delta_t}{Numeric vector.
Vector of time intervals.}
}
\description{
The function generates delta method confidence intervals
for the mediation model \eqn{Y \to M \to X} (Standardized).
}
\examples{
\dontrun{
set.seed(42)
library(dynr)
sim <- GenData(taskid = 1)
data <- RandomMeasurement(sim)
fit <- FitDynr(data, taskid = 1)
theta_hat <- ThetaHat(fit)
ci <- DeltaStdYMX(theta_hat)
plot(ci)
}
}
\seealso{
Other Confidence Interval Functions: 
\code{\link{BootPara}()},
\code{\link{BootParaStdXMY}()},
\code{\link{BootParaStdXYM}()},
\code{\link{BootParaStdYMX}()},
\code{\link{BootParaXMY}()},
\code{\link{BootParaXYM}()},
\code{\link{BootParaYMX}()},
\code{\link{DeltaStdXMY}()},
\code{\link{DeltaStdXYM}()},
\code{\link{DeltaXMY}()},
\code{\link{DeltaXYM}()},
\code{\link{DeltaYMX}()},
\code{\link{IllustrationBootPara}()},
\code{\link{MCStdXMY}()},
\code{\link{MCStdXYM}()},
\code{\link{MCStdYMX}()},
\code{\link{MCXMY}()},
\code{\link{MCXYM}()},
\code{\link{MCYMX}()}
}
\concept{Confidence Interval Functions}
\keyword{ci}
\keyword{manCTMed}
