% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manCTMed-delta-y-m-x.R
\name{DeltaYMX}
\alias{DeltaYMX}
\title{Delta Method Confidence Intervals for Y-M-X}
\usage{
DeltaYMX(phi_hat, delta_t = 1:30)
}
\arguments{
\item{phi_hat}{R object.
Output of the \code{\link[=PhiHat]{PhiHat()}} function.}

\item{delta_t}{Numeric vector.
Vector of time intervals.}
}
\description{
The function generates delta method confidence intervals
for the mediation model \eqn{Y \to M \to X}.
}
\examples{
\dontrun{
set.seed(42)
library(dynr)
sim <- GenData(n = 50)
data <- RandomMeasurement(sim)
fit <- FitDynr(data)
phi_hat <- PhiHat(fit)
ci <- DeltaYMX(phi_hat)
plot(ci)
}
}
\seealso{
Other Confidence Interval Functions: 
\code{\link{DeltaXMY}()},
\code{\link{MCXMY}()},
\code{\link{MCYMX}()},
\code{\link{PhiHat}()}
}
\concept{Confidence Interval Functions}
\keyword{ci}
\keyword{manCTMed}
