% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manCTMed-fig-plot-effects.R
\name{FigPlotEffects}
\alias{FigPlotEffects}
\title{Plot Total, Direct, and Indirect Effects}
\usage{
FigPlotEffects()
}
\description{
Effects for the model \eqn{X \to M \to Y}.
}
\examples{
FigPlotEffects()

}
\seealso{
Other Figure Functions: 
\code{\link{FigScatterPlotCoverage}()},
\code{\link{FigScatterPlotPower}()},
\code{\link{FigScatterPlotType1}()}
}
\author{
Ivan Jacob Agaloos Pesigan

Total, direct, and indirect effects for the drift matrix
\deqn{
    \left(
    \begin{array}{ccc}
         -0.357 & 0 & 0 \\
         0.771 & -0.511 & 0 \\
         -0.450 & 0.729 & -0.693 \\
    \end{array}
    \right)
}
}
\concept{Figure Functions}
\keyword{figure}
\keyword{manCTMed}
