% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manCTMed-fig-plot-effects.R
\name{FigPlotEffects}
\alias{FigPlotEffects}
\title{Plot Total, Direct, and Indirect Effects}
\usage{
FigPlotEffects(dynamics = 0, std = FALSE, max_delta_t = 30, xmy = TRUE)
}
\arguments{
\item{dynamics}{Integer.
\code{dynamics = 0} for original drift matrix,
\code{dynamics = -1} for near-neutral dynamics, and
\code{dynamics = 1} for stronger damping.}

\item{std}{Logical.
If \code{std = TRUE}, standardized total, direct, and indirect effects.
If \code{std = FALSE}, unstandardized total, direct, and indirect effects.}

\item{max_delta_t}{Numeric.
Maximum time interval.}

\item{xmy}{Logical.
If \code{xmy = TRUE},
plot the effects for the \code{x} -> m -> y\verb{mediation model. If}xmy = FALSE\verb{, plot the effects for the }y -> m -> x` mediation model.}
}
\description{
Effects for the model \eqn{X \to M \to Y}.
}
\examples{
FigPlotEffects()

}
\seealso{
Other Figure Functions: 
\code{\link{FigScatterPlotCoverage}()},
\code{\link{FigScatterPlotPower}()},
\code{\link{FigScatterPlotSeBias}()},
\code{\link{FigScatterPlotType1}()},
\code{\link{IllustrationFigPlotEffects}()},
\code{\link{IllustrationFigScatterPlotCoverage}()},
\code{\link{IllustrationFigScatterPlotPower}()},
\code{\link{IllustrationFigScatterPlotSeBias}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Figure Functions}
\keyword{figure}
\keyword{manCTMed}
