% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manCTMed-fig-scatter-plot-coverage.R
\name{FigScatterPlotCoverage}
\alias{FigScatterPlotCoverage}
\title{Plot Coverage Probabilities}
\usage{
FigScatterPlotCoverage(results, delta_t = NULL, R = 20000)
}
\arguments{
\item{results}{Summary results data frame.}

\item{delta_t}{Vector of time-interval value.
If \code{delta_t = NULL}, use all available time-intervals}

\item{R}{Integer.
Number of replications for the Monte Carlo method.
Possible values are 1000, 20000, and 100000.}
}
\description{
Coverage probabilities for the model \eqn{X \to M \to Y}.
}
\examples{
data(results, package = "manCTMed")
FigScatterPlotCoverage(results)
FigScatterPlotCoverage(results, delta_t = 1:14)
FigScatterPlotCoverage(results, delta_t = 15:30)

}
\seealso{
Other Figure Functions: 
\code{\link{FigPlotEffects}()},
\code{\link{FigScatterPlotPower}()},
\code{\link{FigScatterPlotType1}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Figure Functions}
\keyword{figure}
\keyword{manCTMed}
