% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manCTMed-fig-scatter-plot-power.R
\name{FigScatterPlotPower}
\alias{FigScatterPlotPower}
\title{Plot Statistical Power}
\usage{
FigScatterPlotPower(results, delta_t = NULL)
}
\arguments{
\item{results}{Summary results data frame.}

\item{delta_t}{Vector of time-interval value.
If \code{delta_t = NULL}, use all available time-intervals}
}
\description{
Statistical power for the model \eqn{X \to M \to Y}.
}
\examples{
data(results, package = "manCTMed")
FigScatterPlotPower(results)
FigScatterPlotPower(results, delta_t = 1:14)
FigScatterPlotPower(results, delta_t = 15:30)

}
\seealso{
Other Figure Functions: 
\code{\link{FigPlotEffects}()},
\code{\link{FigScatterPlotCoverage}()},
\code{\link{FigScatterPlotType1}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Figure Functions}
\keyword{figure}
\keyword{manCTMed}
