% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manCTMed-fig-scatter-plot-se-bias.R
\name{FigScatterPlotSeBias}
\alias{FigScatterPlotSeBias}
\title{Plot Standard Error Bias}
\usage{
FigScatterPlotSeBias(results, delta_t = NULL, dynamics = 0, std = FALSE)
}
\arguments{
\item{results}{Summary results data frame.}

\item{delta_t}{Vector of time-interval value.
If \code{delta_t = NULL}, use all available time-intervals}

\item{dynamics}{Integer.
\code{dynamics = 0} for original drift matrix,
\code{dynamics = -1} for near-neutral dynamics, and
\code{dynamics = 1} for stronger damping.}

\item{std}{Logical.
If \code{std = TRUE}, standardized total, direct, and indirect effects.
If \code{std = FALSE}, unstandardized total, direct, and indirect effects.}
}
\description{
Standard Error Bias for the model \eqn{X \to M \to Y}.
}
\examples{
data(results, package = "manCTMed")
FigScatterPlotSeBias(results)

}
\seealso{
Other Figure Functions: 
\code{\link{FigPlotEffects}()},
\code{\link{FigScatterPlotCoverage}()},
\code{\link{FigScatterPlotPower}()},
\code{\link{FigScatterPlotType1}()},
\code{\link{IllustrationFigPlotEffects}()},
\code{\link{IllustrationFigScatterPlotCoverage}()},
\code{\link{IllustrationFigScatterPlotPower}()},
\code{\link{IllustrationFigScatterPlotSeBias}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Figure Functions}
\keyword{figure}
\keyword{manCTMed}
