% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manCTMed-fig-scatter-plot-type-1.R
\name{FigScatterPlotType1}
\alias{FigScatterPlotType1}
\title{Plot Type I Error}
\usage{
FigScatterPlotType1(results, delta_t = NULL, R = 20000)
}
\arguments{
\item{results}{Summary results data frame.}

\item{delta_t}{Vector of time-interval value.
If \code{delta_t = NULL}, use all available time-intervals}

\item{R}{Integer.
Number of replications for the Monte Carlo method.
Possible values are 1000, 20000, and 100000.}
}
\description{
Type I error for the model \eqn{Y \to M \to X}.
}
\examples{
data(results, package = "manCTMed")
FigScatterPlotType1(results)
FigScatterPlotType1(results, delta_t = 1:14)
FigScatterPlotType1(results, delta_t = 15:30)

}
\seealso{
Other Figure Functions: 
\code{\link{FigPlotEffects}()},
\code{\link{FigScatterPlotCoverage}()},
\code{\link{FigScatterPlotPower}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Figure Functions}
\keyword{figure}
\keyword{manCTMed}
