% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manCTMed-fig-scatter-plot-type-1.R
\name{FigScatterPlotType1}
\alias{FigScatterPlotType1}
\title{Plot Type I Error}
\usage{
FigScatterPlotType1(results, delta_t = NULL, dynamics = 0, std = FALSE)
}
\arguments{
\item{results}{Summary results data frame.}

\item{delta_t}{Vector of time-interval value.
If \code{delta_t = NULL}, use all available time-intervals}

\item{dynamics}{Integer.
\code{dynamics = 0} for original drift matrix,
\code{dynamics = -1} for near-neutral dynamics, and
\code{dynamics = 1} for stronger damping.}

\item{std}{Logical.
If \code{std = TRUE}, standardized total, direct, and indirect effects.
If \code{std = FALSE}, unstandardized total, direct, and indirect effects.}
}
\description{
Type I error for the model \eqn{Y \to M \to X}.
}
\examples{
data(results, package = "manCTMed")
FigScatterPlotType1(results)
FigScatterPlotType1(results, delta_t = 1:14)
FigScatterPlotType1(results, delta_t = 15:30)

}
\seealso{
Other Figure Functions: 
\code{\link{FigPlotEffects}()},
\code{\link{FigScatterPlotCoverage}()},
\code{\link{FigScatterPlotPower}()},
\code{\link{FigScatterPlotSeBias}()},
\code{\link{IllustrationFigPlotEffects}()},
\code{\link{IllustrationFigScatterPlotCoverage}()},
\code{\link{IllustrationFigScatterPlotPower}()},
\code{\link{IllustrationFigScatterPlotSeBias}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Figure Functions}
\keyword{figure}
\keyword{manCTMed}
