% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manCTMed-fit-dynr.R
\name{FitDynr}
\alias{FitDynr}
\title{Fit the Model using the dynr Package}
\usage{
FitDynr(x)
}
\arguments{
\item{x}{R object.
Output of the \code{\link[=Data]{Data()}} function.}
}
\description{
The function fits the model using the \link{dynr} package.
}
\examples{
library(dynr)
data <- Data(repid = 1, n = 50)
fit <- FitDynr(x = data)

}
\seealso{
Other Simulation Functions: 
\code{\link{CI}()},
\code{\link{Collate}()},
\code{\link{Data}()},
\code{\link{FitCtsem}()},
\code{\link{Replication}()},
\code{\link{Summarize}()}
}
\concept{Simulation Functions}
\keyword{manCTMed}
