% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manCTMed-fit-dynr.R
\name{FitDynr}
\alias{FitDynr}
\title{Fit the Model using the dynr Package}
\usage{
FitDynr(data)
}
\arguments{
\item{data}{R object.
Output of the \code{\link[=RandomMeasurement]{RandomMeasurement()}} function.}
}
\description{
The function fits the model using the \link{dynr} package.
}
\examples{
\dontrun{
set.seed(42)
library(dynr)
sim <- GenData(n = 50)
data <- RandomMeasurement(sim)
FitDynr(data)
}
}
\seealso{
Other Model Fitting Functions: 
\code{\link{PhiHat}()}
}
\concept{Model Fitting Functions}
\keyword{fit}
\keyword{manCTMed}
