% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manCTMed-gen-data.R
\name{GenData}
\alias{GenData}
\title{Simulate Data}
\usage{
GenData(n)
}
\arguments{
\item{n}{Positive integer.
Sample size.}
}
\description{
The function simulates data using
the \code{\link[simStateSpace:SimSSMOUFixed]{simStateSpace::SimSSMOUFixed()}} function.
}
\examples{
\dontrun{
set.seed(42)
sim <- GenData(n = 50)
plot(sim)
}
}
\keyword{gendata}
\keyword{manCTMed}
