% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manCTMed-gen-data.R
\name{GenData}
\alias{GenData}
\title{Simulate Data}
\usage{
GenData(taskid)
}
\arguments{
\item{taskid}{Positive integer.
Task ID.}
}
\description{
The function simulates data using
the \code{\link[simStateSpace:SimSSMOUFixed]{simStateSpace::SimSSMOUFixed()}} function.
}
\examples{
\dontrun{
set.seed(42)
sim <- GenData(taskid = 1)
plot(sim)
}
}
\seealso{
Other Data Generation Functions: 
\code{\link{IllustrationGenData}()},
\code{\link{RandomMeasurement}()}
}
\concept{Data Generation Functions}
\keyword{gendata}
\keyword{manCTMed}
