% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manCTMed-illustration-boot-para.R
\name{IllustrationBootPara}
\alias{IllustrationBootPara}
\title{Parametric Bootstrap (Illustration)}
\usage{
IllustrationBootPara(
  fit,
  path,
  prefix,
  taskid,
  B = 1000L,
  ncores = NULL,
  seed = NULL
)
}
\arguments{
\item{fit}{R object.
Fitted CT-VAR model.}

\item{path}{Path to a directory
to store bootstrap samples and estimates.}

\item{prefix}{Character string.
Prefix used for the file names
for the bootstrap samples and estimates.}

\item{taskid}{Positive integer.
Task ID.}

\item{B}{Positive integer.
Number of bootstrap samples.}

\item{ncores}{Positive integer.
Number of cores to use.}

\item{seed}{Integer.
Random seed.}
}
\description{
The function generates simulated datasets
based on a fitted model and refits the model
to each generated dataset using the \code{dynr} package.
}
\examples{
\dontrun{
library(dynr)
sim <- IllustrationGenData(seed = 42)
data <- IllustrationPrepData(sim)
fit <- IllustrationFitDynr(data)
summary(fit)
IllustrationBootPara(
  fit = fit,
  path = getwd(),
  prefix = "pb",
  taskid = 1,
  B = 1000L,
  seed = 42
)
}
}
\seealso{
Other Confidence Interval Functions: 
\code{\link{BootPara}()},
\code{\link{BootParaStdXMY}()},
\code{\link{BootParaStdXYM}()},
\code{\link{BootParaStdYMX}()},
\code{\link{BootParaXMY}()},
\code{\link{BootParaXYM}()},
\code{\link{BootParaYMX}()},
\code{\link{DeltaStdXMY}()},
\code{\link{DeltaStdXYM}()},
\code{\link{DeltaStdYMX}()},
\code{\link{DeltaXMY}()},
\code{\link{DeltaXYM}()},
\code{\link{DeltaYMX}()},
\code{\link{MCStdXMY}()},
\code{\link{MCStdXYM}()},
\code{\link{MCStdYMX}()},
\code{\link{MCXMY}()},
\code{\link{MCXYM}()},
\code{\link{MCYMX}()}
}
\concept{Confidence Interval Functions}
\keyword{ci}
\keyword{illustration}
\keyword{manCTMed}
