% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manCTMed-illustration-fig-plot-effects.R
\name{IllustrationFigPlotEffects}
\alias{IllustrationFigPlotEffects}
\title{Plot Total, Direct, and Indirect Effects (Illustration)}
\usage{
IllustrationFigPlotEffects(std = FALSE, max_delta_t = 30)
}
\arguments{
\item{std}{Logical.
If \code{std = TRUE}, standardized total, direct, and indirect effects.
If \code{std = FALSE}, unstandardized total, direct, and indirect effects.}

\item{max_delta_t}{Numeric.
Maximum time interval.}
}
\description{
Effects for the model \eqn{X \to M \to Y}.
}
\examples{
IllustrationFigPlotEffects(std = FALSE)
IllustrationFigPlotEffects(std = TRUE)

}
\seealso{
Other Figure Functions: 
\code{\link{FigPlotEffects}()},
\code{\link{FigScatterPlotCoverage}()},
\code{\link{FigScatterPlotPower}()},
\code{\link{FigScatterPlotSeBias}()},
\code{\link{FigScatterPlotType1}()},
\code{\link{IllustrationFigScatterPlotCoverage}()},
\code{\link{IllustrationFigScatterPlotPower}()},
\code{\link{IllustrationFigScatterPlotSeBias}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Figure Functions}
\keyword{figure}
\keyword{manCTMed}
