% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/manCTMed-illustration-fig-scatter-plot-coverage.R
\name{IllustrationFigScatterPlotCoverage}
\alias{IllustrationFigScatterPlotCoverage}
\title{Illustration Plot Coverage Probabilities}
\usage{
IllustrationFigScatterPlotCoverage(illustration_results)
}
\arguments{
\item{illustration_results}{Summary results data frame.}
}
\description{
Coverage probabilities for the model \eqn{X \to M \to Y}.
}
\examples{
data(illustration_results, package = "manCTMed")
IllustrationFigScatterPlotCoverage(illustration_results)

}
\seealso{
Other Figure Functions: 
\code{\link{FigPlotEffects}()},
\code{\link{FigScatterPlotCoverage}()},
\code{\link{FigScatterPlotPower}()},
\code{\link{FigScatterPlotSeBias}()},
\code{\link{FigScatterPlotType1}()},
\code{\link{IllustrationFigPlotEffects}()},
\code{\link{IllustrationFigScatterPlotPower}()},
\code{\link{IllustrationFigScatterPlotSeBias}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Figure Functions}
\keyword{figure}
\keyword{illustration}
\keyword{manCTMed}
