% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/manCTMed-illustration-fig-scatter-plot-se-bias.R
\name{IllustrationFigScatterPlotSeBias}
\alias{IllustrationFigScatterPlotSeBias}
\title{Illustration Plot Standard Error Bias}
\usage{
IllustrationFigScatterPlotSeBias(illustration_results)
}
\arguments{
\item{illustration_results}{Summary results data frame.}
}
\description{
Standard Error Bias for the model \eqn{X \to M \to Y}.
}
\examples{
data(illustration_results, package = "manCTMed")
IllustrationFigScatterPlotSeBias(illustration_results)

}
\seealso{
Other Figure Functions: 
\code{\link{FigPlotEffects}()},
\code{\link{FigScatterPlotCoverage}()},
\code{\link{FigScatterPlotPower}()},
\code{\link{FigScatterPlotSeBias}()},
\code{\link{FigScatterPlotType1}()},
\code{\link{IllustrationFigPlotEffects}()},
\code{\link{IllustrationFigScatterPlotCoverage}()},
\code{\link{IllustrationFigScatterPlotPower}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Figure Functions}
\keyword{figure}
\keyword{illustration}
\keyword{manCTMed}
