% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manCTMed-illustration-gen-data.R
\name{IllustrationGenData}
\alias{IllustrationGenData}
\title{Simulate Data (Illustration)}
\usage{
IllustrationGenData(seed = NULL, n = 133, m = 101, delta_t_gen = 0.1)
}
\arguments{
\item{seed}{Integer.
Random seed.}

\item{n}{Positive integer.
Sample size.}

\item{m}{Positive integer.
Measurement occasions.}

\item{delta_t_gen}{Numeric.
Time interval used to generate data.}
}
\description{
The function simulates data using
the \code{\link[simStateSpace:SimSSMOUFixed]{simStateSpace::SimSSMOUFixed()}} function.
}
\examples{
\dontrun{
sim <- IllustrationGenData(seed = 42)
plot(sim)
}
}
\seealso{
Other Data Generation Functions: 
\code{\link{GenData}()},
\code{\link{RandomMeasurement}()}
}
\concept{Data Generation Functions}
\keyword{gendata}
\keyword{illustration}
\keyword{manCTMed}
