% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manCTMed-illustration-prep-data.R
\name{IllustrationPrepData}
\alias{IllustrationPrepData}
\title{Prepare Data Before Model Fitting (Illustration)}
\usage{
IllustrationPrepData(sim)
}
\arguments{
\item{sim}{R object.
Output of the \code{\link[=IllustrationGenData]{IllustrationGenData()}} function.}
}
\description{
The function converts the output of \code{\link[=IllustrationGenData]{IllustrationGenData()}}
into a data frame.
}
\examples{
\dontrun{
sim <- IllustrationGenData(seed = 42)
data <- IllustrationPrepData(sim)
head(data)
dim(data)
}
}
\seealso{
Other Model Fitting Functions: 
\code{\link{FitDynr}()},
\code{\link{FitMx}()},
\code{\link{IllustrationFitDynr}()},
\code{\link{IllustrationFitMx}()},
\code{\link{IllustrationMCPhiSigma}()},
\code{\link{PhiHat}()},
\code{\link{ThetaHat}()}
}
\concept{Model Fitting Functions}
\keyword{gendata}
\keyword{illustration}
\keyword{manCTMed}
