% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manCTMed-package.R
\docType{package}
\name{manCTMed-package}
\alias{manCTMed}
\alias{manCTMed-package}
\title{manCTMed: Continuous Time Mediation}
\description{
Research compendium for the manuscript Pesigan, I. J. A., Russell, M. A., Chow, S.-M. (In Preparation). Standard Errors and Confidence Intervals for Total, Direct, and Indirect Effects in Continuous-Time Mediation Models. \doi{10.0000/0000000000}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jeksterslab/manCTMed}
  \item \url{https://jeksterslab.github.io/manCTMed/}
  \item \url{https://osf.io/qwnmf/?view_only=855be2012b554f05bbc679946845bab8}
  \item \doi{10.0000/0000000000}
  \item Report bugs at \url{https://github.com/jeksterslab/manCTMed/issues}
}

}
\author{
\strong{Maintainer}: Ivan Jacob Agaloos Pesigan \email{r.jeksterslab@gmail.com} (\href{https://orcid.org/0000-0000-0000-0000}{ORCID}) [copyright holder]

}
\keyword{internal}
