% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manCTMed-mc-std-x-m-y.R
\name{MCStdXMY}
\alias{MCStdXMY}
\title{Monte Carlo Method Confidence Intervals for X-M-Y (Standardized)}
\usage{
MCStdXMY(theta_hat, delta_t = 1:30, R = 20000L, seed = NULL)
}
\arguments{
\item{theta_hat}{R object.
Output of the \code{\link[=ThetaHat]{ThetaHat()}} function.}

\item{delta_t}{Numeric vector.
Vector of time intervals.}

\item{R}{Positive integer.
Number of random drift matrices to generate.}

\item{seed}{Integer.
Random seed.}
}
\description{
The function generates Monte Carlo method confidence intervals
for the mediation model \eqn{X \to M \to Y} (Standardized).
}
\examples{
\dontrun{
set.seed(42)
library(dynr)
sim <- GenData(taskid = 1)
data <- RandomMeasurement(sim)
fit <- FitDynr(data, taskid = 1)
theta_hat <- ThetaHat(fit)
ci <- MCStdXMY(theta_hat, seed = 42)
plot(ci)
}
}
\seealso{
Other Confidence Interval Functions: 
\code{\link{BootPara}()},
\code{\link{BootParaStdXMY}()},
\code{\link{BootParaStdYMX}()},
\code{\link{BootParaXMY}()},
\code{\link{BootParaYMX}()},
\code{\link{DeltaStdXMY}()},
\code{\link{DeltaStdYMX}()},
\code{\link{DeltaXMY}()},
\code{\link{DeltaYMX}()},
\code{\link{MCStdYMX}()},
\code{\link{MCXMY}()},
\code{\link{MCYMX}()}
}
\concept{Confidence Interval Functions}
\keyword{ci}
\keyword{manCTMed}
