% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manCTMed-mc-x-m-y.R
\name{MCXMY}
\alias{MCXMY}
\title{Monte Carlo Method Confidence Intervals for X-M-Y}
\usage{
MCXMY(phi_hat, delta_t = 1:30, R = 20000L, seed)
}
\arguments{
\item{phi_hat}{R object.
Output of the \code{\link[=PhiHat]{PhiHat()}} function.}

\item{delta_t}{Numeric vector.
Vector of time intervals.}

\item{R}{Positive integer.
Number of random drift matrices to generate.}

\item{seed}{Integer.
Random seed.}
}
\description{
The function generates Monte Carlo method confidence intervals
for the mediation model \eqn{X \to M \to Y}.
}
\examples{
\dontrun{
set.seed(42)
library(dynr)
sim <- GenData(n = 50)
data <- RandomMeasurement(sim)
fit <- FitDynr(data)
phi_hat <- PhiHat(fit)
ci <- MCXMY(phi_hat)
plot(ci)
}
}
\keyword{ci}
\keyword{manCTMed}
