% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manCTMed-mc-x-m-y.R
\name{MCXMY}
\alias{MCXMY}
\title{Monte Carlo Method Confidence Intervals for X-M-Y}
\usage{
MCXMY(phi_hat, delta_t = 1:30, R = 20000L, seed = NULL)
}
\arguments{
\item{phi_hat}{R object.
Output of the \code{\link[=PhiHat]{PhiHat()}} function.}

\item{delta_t}{Numeric vector.
Vector of time intervals.}

\item{R}{Positive integer.
Number of Monte Carlo replications.}

\item{seed}{Integer.
Random seed.}
}
\description{
The function generates Monte Carlo method confidence intervals
for the mediation model \eqn{X \to M \to Y}.
}
\examples{
\dontrun{
set.seed(42)
library(dynr)
sim <- GenData(taskid = 1)
data <- RandomMeasurement(sim)
fit <- FitDynr(data, taskid = 1)
phi_hat <- PhiHat(fit)
ci <- MCXMY(phi_hat, seed = 42)
plot(ci)
}
}
\seealso{
Other Confidence Interval Functions: 
\code{\link{BootPara}()},
\code{\link{BootParaStdXMY}()},
\code{\link{BootParaStdXYM}()},
\code{\link{BootParaStdYMX}()},
\code{\link{BootParaXMY}()},
\code{\link{BootParaXYM}()},
\code{\link{BootParaYMX}()},
\code{\link{DeltaStdXMY}()},
\code{\link{DeltaStdXYM}()},
\code{\link{DeltaStdYMX}()},
\code{\link{DeltaXMY}()},
\code{\link{DeltaXYM}()},
\code{\link{DeltaYMX}()},
\code{\link{IllustrationBootPara}()},
\code{\link{MCStdXMY}()},
\code{\link{MCStdXYM}()},
\code{\link{MCStdYMX}()},
\code{\link{MCXYM}()},
\code{\link{MCYMX}()}
}
\concept{Confidence Interval Functions}
\keyword{ci}
\keyword{manCTMed}
