% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manCTMed-data-params.R
\docType{data}
\name{params}
\alias{params}
\title{Simulation Parameters}
\format{
A dataframe with 30 rows and 3 columns:

\describe{
\item{taskid}{
Simulation Task ID.
}
\item{n}{
Sample size.
}
\item{dynamics}{
Dynamics.
\code{0} for original drift matrix,
\code{-1} for near-neutral dynamics, and
\code{1} for stronger damping.
}
}
}
\usage{
data(params)
}
\description{
Simulation Parameters
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{data}
\keyword{parameters}
