% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manCTMed-phi-hat.R
\name{PhiHat}
\alias{PhiHat}
\title{Estimated Drift Matrix}
\usage{
PhiHat(fit)
}
\arguments{
\item{fit}{R object.
Output of the \code{\link[=FitDynr]{FitDynr()}} function.}
}
\description{
The function extracts the estimated drift matrix
from the fitted model.
}
\examples{
\dontrun{
set.seed(42)
library(dynr)
sim <- GenData(n = 50)
data <- RandomMeasurement(sim)
fit <- FitDynr(data)
PhiHat(fit)
}
}
\seealso{
Other Confidence Interval Functions: 
\code{\link{DeltaXMY}()},
\code{\link{DeltaYMX}()},
\code{\link{MCXMY}()},
\code{\link{MCYMX}()}
}
\concept{Confidence Interval Functions}
\keyword{ci}
\keyword{manCTMed}
