% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manCTMed-plot-coverage.R
\name{PlotCoverage}
\alias{PlotCoverage}
\title{Plot Coverage Probabilities}
\usage{
PlotCoverage(x)
}
\arguments{
\item{x}{R object.
Output of the \code{\link[=Summarize]{Summarize()}} function.}
}
\description{
Plot Coverage Probabilities
}
\seealso{
Other Simulation Functions: 
\code{\link{CI}()},
\code{\link{Collate}()},
\code{\link{Data}()},
\code{\link{FitDynr}()},
\code{\link{PlotType1}()},
\code{\link{Replication}()},
\code{\link{Summarize}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Simulation Functions}
\keyword{manCTMed}
