% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manCTMed-plot-type-1.R
\name{PlotType1}
\alias{PlotType1}
\title{Plot Type I Error}
\usage{
PlotType1(x)
}
\arguments{
\item{x}{Summary results data frame.}
}
\description{
Plot Type I Error
}
\seealso{
Other Simulation Functions: 
\code{\link{PlotCoverage}()},
\code{\link{PlotPower}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Simulation Functions}
\keyword{manCTMed}
