% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manCTMed-replication.R
\name{Replication}
\alias{Replication}
\title{Simulation Replication}
\usage{
Replication(i, n, write = FALSE, wd)
}
\arguments{
\item{i}{Positive integer.
Replication ID.}

\item{n}{Positive integer.
Sample size.}

\item{write}{Logical.
If \code{write = TRUE}, write results into a file.
If \code{write = FALSE}, return results.}

\item{wd}{Working directory.
Directory where output file is saved
if \code{write = TRUE}.}
}
\description{
Performs a single simulation replication involving the following:
\itemize{
\item Data generation using the \code{simStateSpace} package.
\item Model fitting using the \code{dynr} package.
\item Model fitting using the \code{ctsem} package.
\item Generate delta method confidence intervals
using the \code{\link[cTMed:DeltaMed]{cTMed::DeltaMed()}} function.
\item Generate Monte Carlo method confidence intervals
using the \code{\link[cTMed:MCMed]{cTMed::MCMed()}} function.
\item Generate credible intervals
based on the posterior distribution of the drift matrix
using the \code{\link[cTMed:PosteriorMed]{cTMed::PosteriorMed()}} function.
}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Simulation Functions}
\keyword{manCTMed}
