% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manCTMed-replication.R
\name{Replication}
\alias{Replication}
\title{Simulation Replication}
\usage{
Replication(repid, n, wd, delta_t = 1:24, R = 20000L)
}
\arguments{
\item{repid}{Positive integer.
Replication ID.}

\item{n}{Positive integer.
Sample size.}

\item{wd}{Working directory for output files.}

\item{delta_t}{Numeric vector.
Vector of time-intervals to use.}

\item{R}{Positive integer.
Number of Monte Carlo replications.}
}
\description{
Performs a single simulation replication involving the following:
\itemize{
\item Data generation using the \code{simStateSpace} package.
\item Model fitting using the \code{dynr} package.
\item Generate delta method confidence intervals
using the \code{\link[cTMed:DeltaMed]{cTMed::DeltaMed()}} function.
\item Generate Monte Carlo method confidence intervals
using the \code{\link[cTMed:MCMed]{cTMed::MCMed()}} function.
}
}
\seealso{
Other Simulation Functions: 
\code{\link{CI}()},
\code{\link{Collate}()},
\code{\link{Data}()},
\code{\link{FitDynr}()},
\code{\link{PlotCoverage}()},
\code{\link{PlotType1}()},
\code{\link{Summarize}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Simulation Functions}
\keyword{manCTMed}
