% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manCTMed-sim.R
\name{Sim}
\alias{Sim}
\title{Simulation Replication}
\usage{
Sim(
  taskid,
  repid,
  output_folder,
  overwrite,
  integrity,
  delta_t,
  R,
  B,
  run_boot = FALSE
)
}
\arguments{
\item{taskid}{Positive integer.
Task ID.}

\item{repid}{Positive integer.
Replication ID.}

\item{output_folder}{Character string.
Output folder.}

\item{overwrite}{Logical.
Overwrite existing output in \code{output_folder}.}

\item{integrity}{Logical.
If \code{integrity = TRUE},
check for the output file integrity when \code{overwrite = FALSE}.}

\item{delta_t}{Numeric vector.
Vector of time intervals.}

\item{R}{Positive integer.
Number of random drift matrices to generate.}

\item{B}{Positive integer.
Number of bootstrap samples.}

\item{run_boot}{Logical.
If \code{run_boot = TRUE}, run bootstrap.}
}
\value{
The output is saved as an external file in \code{output_folder}.
}
\description{
Simulation Replication
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{manCTMed}
\keyword{simulation}
