% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manCTMed-summarize.R
\name{Summarize}
\alias{Summarize}
\title{Summarize Results}
\usage{
Summarize(x, ncores = NULL)
}
\arguments{
\item{x}{R object.
Output of the \code{\link[=Collate]{Collate()}} function.}

\item{ncores}{Positive integer.
Number of cores to use.}
}
\description{
Summarize Results
}
\seealso{
Other Simulation Functions: 
\code{\link{CI}()},
\code{\link{Collate}()},
\code{\link{Data}()},
\code{\link{FitDynr}()},
\code{\link{PlotCoverage}()},
\code{\link{PlotType1}()},
\code{\link{Replication}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Simulation Functions}
\keyword{manCTMed}
