% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manCTMed-theta-hat.R
\name{ThetaHat}
\alias{ThetaHat}
\title{Estimated Drift Matrix and Process Noise}
\usage{
ThetaHat(fit)
}
\arguments{
\item{fit}{R object.
Output of the \code{\link[=FitDynr]{FitDynr()}}, \code{\link[=FitMx]{FitMx()}},
\code{\link[=IllustrationFitDynr]{IllustrationFitDynr()}}, or \code{\link[=IllustrationFitMx]{IllustrationFitMx()}},
functions.}
}
\description{
The function extracts the estimated drift matrix and process noise
from the fitted model.
}
\examples{
\dontrun{
set.seed(42)
library(dynr)
sim <- GenData(n = 50)
data <- RandomMeasurement(sim)
fit <- FitDynr(data)
ThetaHat(fit)
}
}
\seealso{
Other Model Fitting Functions: 
\code{\link{FitDynr}()},
\code{\link{FitMx}()},
\code{\link{IllustrationFitDynr}()},
\code{\link{IllustrationFitMx}()},
\code{\link{IllustrationMCPhiSigma}()},
\code{\link{IllustrationPrepData}()},
\code{\link{PhiHat}()}
}
\concept{Model Fitting Functions}
\keyword{ci}
\keyword{manCTMed}
